third_party/blink/web_tests/external/wpt/html/semantics/the-button-element/command-and-commandfor/event-dispatch-shadow-expected.txt

This is a testharness.js-based test.
[FAIL] cross shadow CommandEvent retargets source to host element
  assert_equals: composed is false expected false but got true
Harness: the test ran to completion.
